Break main into smaller functions, move includes into root.h.
The main function was becoming unweildy, largely due to error
handling and the absolute/qualified/unqualified path logic.
I turned each logical step into a function and called it from
main, so that main is now only a top-level overview.

On the (possible) downside, now lots of functions can call
exit().  The current intent is that only functions in the "main"
module should call exit().

As noted, all the header include and define mess was moved into
root.h.  This is based on comments by Rob Pike in his books and
articles.  If root.c ever defines a public API, this decision
may need to be re-assessed.